Added the new evaluation.py, please install tshark and wireshark-common packages.

+++ evaluation.py
@@ -1,7 +1,10 @@
+import logging
+logging.getLogger("scapy.runtime").setLevel(logging.ERROR)
 import os
 from scapy.all import *
 import re
 import time
+from datetime import datetime
 
 num_host = 16
 
@@ -32,6 +35,7 @@ def traffic_spec(host,hostip):
             traffic[hostip+':'+'10.0.0.'+dstid+':'+dstport] = vol_num
     return traffic
 
+
     
 #compare dump file with traffic spec
 def get_dump_stat(dump_file, spec, hostip):
@@ -40,45 +44,106 @@ def get_dump_stat(dump_file, spec, hostip):
     #global g_correctness_list
     global g_stat
 
-    pkttrace = rdpcap(dump_file)
-    for pkt in pkttrace:
+    p = re.compile("(\S*)\:(\S*)\:(\S*)")
+    for flow in spec: 
+        m = p.match(flow)
+        if m != None:
+            src = m.group(1)
+            dst = m.group(2)
+            port = m.group(3)
+            #print src,dst,port
+            #print dst,hostip
+            if dst == hostip:
+                print "Analyzing trace for flow: " + src + " -> " + dst +":"+port
+                cmd = 'sudo tcpdump vlan and dst ' + dst + ' and src ' + src + ' and dst port ' + port + ' -r ' + dump_file + ' -w dump/tmp.pcap > /dev/null 2>&1'
+                #print cmd
+                os.system(cmd)
+
+                cmd = 'capinfos -Tm dump/tmp.pcap'
+                #print cmd
+                info = os.popen(cmd).read()
+                pos1 = info.find("\n")
+                #print info[:pos1]
+                pos2 = info.find("\n", pos1+1)
+                info_list = info[pos1+1:pos2].split(",")                    
+                #print info_list
+                pkt = int(info_list[6])
+                
+                #here we only calculate payload size
+                #Ether(14) + Vlantag(4) + IP (20) + UDP(8) = 46 Byte header
+                #let us know if this calculation does not work for you
+                byt = int(info_list[8]) - pkt * 46  #total bytes
+                thr = float(info_list[13])       #bps
+                pktrate = float(info_list[15])   #packets/sec
+                pktsize = float(info_list[14])   #average packet size
+                #print pkt, byt, thr,pktrate, pktsize
+               
+                
+                # the duration in capinfos is not precise
+                # instead we use..
+                if pkt == 0:
+                    g_stat[flow] = (byt, 0)
+                else:
+                    cmd = 'tshark -r dump/tmp.pcap -Y "frame.number == 1" -T fields -e frame.time'
+                    #print cmd
+                    start = os.popen(cmd).read()
+                    cmd = 'tshark -r dump/tmp.pcap -Y "frame.number == ' + str(pkt) + '\" -T fields -e frame.time'
+                    #print cmd
+                    end = os.popen(cmd).read()
+
+                    starttime = datetime.strptime(start.split('.')[0] + '.' + start.split('.')[1][0:6],"%b  %d, %Y %H:%M:%S.%f")
+                    endtime = datetime.strptime(end.split('.')[0] + '.' + end.split('.')[1][0:6],"%b  %d, %Y %H:%M:%S.%f")
+                    
+                    
+                    #duration = (endtime-starttime).seconds*1.0 + (endtime-starttime).microseconds * 10.0/1000000
+
+                    epoch = datetime.utcfromtimestamp(0)
+                    delta = (endtime - epoch).total_seconds()
+                    
+                    g_stat[flow] = (byt, delta)
+                    
+                #exit(1) 
+
+    #return 
+    
+    #for pkt in pkttrace:
         #print pkt.time
         #exit(1)
         
-        if pkt.haslayer(IP) == 0 or  pkt.haslayer(UDP) == 0:
-            continue
+     #   if pkt.haslayer(IP) == 0 or  pkt.haslayer(UDP) == 0:
+     #      continue
 
 
-        src = pkt.getlayer(IP).src
-        dst = pkt.getlayer(IP).dst
-        if dst != hostip:
-            continue
+     #   src = pkt.getlayer(IP).src
+     #   dst = pkt.getlayer(IP).dst
+     #   if dst != hostip:
+     #       continue
             
-        dport = pkt.getlayer(IP).dport
-        key = str(src)+':'+str(dst)+':'+str(dport)
-        ts = pkt.time
+        #dport = pkt.getlayer(IP).dport
+        #key = str(src)+':'+str(dst)+':'+str(dport)
+        #ts = pkt.time
         
-        if key not in spec:
-            continue
+        #if key not in spec:
+        #    continue
             
 
-        if key not in g_stat:
-            g_stat[key] = {}
+        #if key not in g_stat:
+        #    g_stat[key] = {}
             
         # suppose you have only ONE 802.1Q layer
-        if pkt.haslayer(Dot1Q):
-            vlan = int(pkt.getlayer(Dot1Q).vlan)
+        #if pkt.haslayer(Dot1Q):
+         #   vlan = int(pkt.getlayer(Dot1Q).vlan)
             #print vlan
 
-        if vlan not in g_stat[key]:
-            g_stat[key][vlan] = (0,ts)
+        #if vlan not in g_stat[key]:
+        #    g_stat[key][vlan] = (0,ts)
             
-        header_size = 20 # IP header
-        if pkt.proto == 17 or pkt.proto == 6:
-            header_size += 8  # L4 header
+        #header_size = 20 # IP header
+        #if pkt.proto == 17 or pkt.proto == 6:
+        #    header_size += 8  # L4 header
 
             
-        g_stat[key][vlan] = (g_stat[key][vlan][0] + (pkt.len - header_size), ts)
+        #g_stat[key][vlan] = (g_stat[key][vlan][0] + (pkt.len - header_size), ts)
 
 
 def stat_summary(spec):
@@ -91,14 +156,16 @@ def stat_summary(spec):
             continue
         flowsize = 0
         finish_time = 0
-        for vlan in g_stat[key]:
+        #for vlan in g_stat[key]:
             #print stat[key][vlan]
-            flowsize += g_stat[key][vlan][0]
+            #flowsize += g_stat[key][vlan][0]
             
-            if g_stat[key][vlan][1] > finish_time:
-                finish_time = g_stat[key][vlan][1]
+            #if g_stat[key][vlan][1] > finish_time:
+            #    finish_time = g_stat[key][vlan][1]
             #test
             #print "vlan: " + key+'\t'+ str(vlan) + '\t' + str(stat[key][vlan])
+        flowsize = g_stat[key][0]
+        finish_time = g_stat[key][1]
         if spec[key] == flowsize:
             result = "OK"
             g_correctness_list.append(1)
@@ -106,8 +173,14 @@ def stat_summary(spec):
             result = "Wrong flow size"
             g_correctness_list.append(0)
 
-        FCT = finish_time - g_start_time
-        throughput = flowsize * 8.0 / (FCT)/ 1e6
+            
+        if finish_time != 0:
+            FCT = finish_time - g_start_time
+            throughput = flowsize * 8.0 / (FCT)/ 1e6
+        else:
+            FCT = float('inf')
+            throughput = 0
+            
     
         print "flow " + key + '\t' + "FCT " + str(FCT) + '\t' + "thr " + str(throughput) +" Mbps\texpt " + str(spec[key]) + '\t' + "dump " + str(flowsize) + '\t' + result
         
@@ -132,14 +205,17 @@ def dump_trace_analysis():
                 traffic[item] = traffic_[item]
             else:
                 traffic[item] += traffic_[item]
-        
+
+    #print traffic
+    
     for hostid in range(1,num_host+1):
         host = 'h' + str(hostid)
         hostip = '10.0.0.'+str(hostid)
         tracefile = 'dump/' + host + '.pcap'
+        #print "process dump for ",host
         get_dump_stat(tracefile, traffic, hostip)
 
-        
+
     #analyze the traffic statistics
     stat_summary(traffic)
     
@@ -154,32 +230,54 @@ def dump_trace_analysis():
     if len(g_FCT_list) == 0:
         print "tail FCT: N/A"
     else:
-        print "tail FCT: " + str(min(g_FCT_list))
+        print "tail FCT: " + str(max(g_FCT_list))
     
 
     
 def kill_all_task():
-    cmd = "ps -ef | grep \'tcpdump\|cperf\' | grep -v grep | awk \'{print $2}\'"
+    _digit = re.compile('\d')
+    cmd = "ps -ef | grep \'tcpdump\' | grep -v grep | awk \'{print $2}\'"
     pids = os.popen(cmd).read()
     pid_all = ""
     strs = pids.split('\n')
     for pid in strs:
         pid_all += str(pid)
         pid_all += " "
-    #print pid_all
-    os.system("sudo kill " + pid_all)
+        
+    #print "pid", pid_all
+    if bool(_digit.search(pid_all)):
+        os.system("sudo kill " + pid_all)
+
+    
+    cmd = "ps -ef | grep \'cperf\|./arbiter\' | grep -v grep | awk \'{print $2}\'"
+    pids = os.popen(cmd).read()
+    pid_all = ""
+    strs = pids.split('\n')
+    for pid in strs:
+        pid_all += str(pid)
+        pid_all += " "
+    #print "pid",pid_all
+    if bool(_digit.search(pid_all)):
+        os.system("sudo kill -9 " + pid_all)
+
 
 if __name__=='__main__':
 
+    print "Clear existing task..."
+    kill_all_task()
+
+    if not os.path.exists("dump"):
+        os.makedirs("dump")
+    
     print "Starting Arbiter..."
-    cmd = "./arbiter"
+    cmd = "./arbiter >/dev/null 2>&1 &"
     os.system(cmd)
     
     print "Starting tcpdump..."
     time.sleep(1)
     for hostid in range(1,num_host+1): 
         host = 'h' + str(hostid)
-        cmd = 'mininet/util/m ' + host + ' sudo tcpdump -i ' + host +'-eth0' + ' -n -s 64 portrange 5000-5005 -w dump/' + host + '.pcap &'
+        cmd = '~/mininet/util/m ' + host + ' sudo tcpdump -i ' + host +'-eth0' + ' -n -s 64 -B 8192 -w dump/' + host + '.pcap >/dev/null 2>&1 &'
         print cmd
         os.system(cmd)
 
@@ -187,16 +285,19 @@ if __name__=='__main__':
     time.sleep(1)
     for hostid in range(1,num_host+1): 
         host = 'h' + str(hostid)
-        cmd = 'mininet/util/m ' + host + ' ./cperf trace/' + host + '.tr &'
+        cmd = '~/mininet/util/m ' + host + ' ./cperf traffic/' + host + '.tr >/dev/null 2>&1 &'
         print cmd
         os.system(cmd)
 
     #sleep sufficient amount of time
     #change the value in need
-    time.sleep(60)
+    print "wait for all flows finished..."
+    time.sleep(120)
 
     print "Terminating all task..."
     kill_all_task()
 
     print "Analyzing traffic statistics..."
     dump_trace_analysis()
+
+
